+++ packages/devnode/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @provablehq/veil-aleo-devnode
 
+## 0.5.0
+
+### Minor Changes
+
+- Version alignment with the 0.5.0 release of the fixed Veil package group
+  (agent skills + DEX API auth in `@provablehq/shield-swap-sdk`, FeeMaster
+  fee payment in `@provablehq/veil-aleo-sdk`).
+
 ## 0.4.1
 
 ### Patch Changes