+++ packages/leo/CHANGELOG.md
@@ -1,3 +1,11 @@
 # @provablehq/veil-leo
 
+## 0.5.0
+
+### Minor Changes
+
+- Version alignment with the 0.5.0 release of the fixed Veil package group
+  (agent skills + DEX API auth in `@provablehq/shield-swap-sdk`, FeeMaster
+  fee payment in `@provablehq/veil-aleo-sdk`).
+
 ## 0.4.1